HPCC-30051 Setup Core Check
The Setup Core Check subtask is part of the improve error handling and reporting
for automated test systems.

Setup Core Check contains the CoreScript.sh which when called upon will go
and detect core files generated during a crash and will create a trace file
out of it with the necessary information for debugging.

Smoketest.yml is also updated to call on CoreScript.sh

Testing of the script was done by using provided core files by Attila
Integration of CoreScript and updates to Smoketest.yml are also
successfully ran through in GitHub Actions.

Signed-off-by: Johnny <johnnyhuang632402@gmail.com>