HPCC-30053 Regression Testing
The Regression Testing subtask is part of
the improve error handling and reporting for automated test systems.

Regression Testing contains the MatrixCoreScript.sh which when called upon
will go and detect core files generated during regression testing.
The resulting core files will be used in the script
to create a trace file out of it with the necessary information for debugging.

Smoketest.yml is also updated to call on MatrixCoreScript.sh

Testing of the script was done by using provided core files by Attila
Integration of MatrixCoreScript and updates to Smoketest.yml
are also successfully ran through in GitHub Actions.

Signed-off-by: Johnny <johnnyhuang632402@gmail.com>